cleaning up directory structure -- test cases for all lab value labelers
